+++ app/Http/Controllers/Api/ConsultationExamenCliniqueController.php
@@ -16,6 +16,12 @@ class ConsultationExamenCliniqueController extends Controller
         ]);
 
         $consultation = ConsultationMedecineGenerale::find($request->get('consultation'));
+
+        $isAuthor = checkIfIsAuthorOrIsAuthorized("ConsultationExamenClinique",$consultation->id,"attach");
+        if($isAuthor->getOriginalContent() == false){
+            return response()->json(['error'=>"Vous ne pouvez modifié un élement que vous n'avez crée"],401);
+        }
+
         $consultation->examensClinique()->detach($request->get('examensClinique'));
 
         $consultation = ConsultationMedecineGenerale::with('examensClinique')->find($request->get('consultation'));
