+++ app/Http/Controllers/Api/ConsultationObstetriqueController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Requests\ConsultationObstetriqueRequest;
 use App\Models\ConsultationObstetrique;
+use Carbon\Carbon;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
@@ -45,7 +46,9 @@ class ConsultationObstetriqueController extends Controller
             $praticen = $user->praticien;
             if ($praticen->specialite->name == "Gynéco-obstétrique"){
                 $consultationObstetrique =  ConsultationObstetrique::create($request->validated());
+
                 defineAsAuthor("ConsultationObstetrique",$consultationObstetrique->id,'create');
+
                 return response()->json(['consultationObstetrique'=>$consultationObstetrique]);
             }else{
                 return response()->json(['error'=>"Vous ne disposez pas des autorisations pour effectuer cette action"],422);
@@ -98,6 +101,11 @@ class ConsultationObstetriqueController extends Controller
         if(!is_null($validation))
             return $validation;
 
+        $isAuthor = checkIfIsAuthorOrIsAuthorized("ConsultationObstetrique",$id,"create");
+        if($isAuthor->getOriginalContent() == false){
+            return response()->json(['error'=>"Vous ne pouvez modifié un élement que vous n'avez crée"],401);
+        }
+
         ConsultationObstetrique::whereId($id)->update($request->validated());
         $consultationObstetrique =  ConsultationObstetrique::find($id);
         return response()->json(['consultationObstetrique'=>$consultationObstetrique]);
@@ -116,7 +124,57 @@ class ConsultationObstetriqueController extends Controller
             return $validation;
 
         $consultationObstetrique =  ConsultationObstetrique::find($id);
+
+        $isAuthor = checkIfIsAuthorOrIsAuthorized("ConsultationObstetrique",$id,"create");
+        if($isAuthor->getOriginalContent() == false){
+            return response()->json(['error'=>"Vous ne pouvez modifié un élement que vous n'avez crée"],401);
+        }
+
         ConsultationObstetrique::destroy($id);
         return response()->json(['consultationObstetrique'=>$consultationObstetrique]);
+    }
+
+    /**
+     * Archieved the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function archiver($id)
+    {
+        $validation = validatedId($id,$this->table);
+        if(!is_null($validation))
+            return $validation;
+
+        $resultat = ConsultationObstetrique::with(['dossier','consultation'])->find($id);
+        if (is_null($resultat->passed_at)){
+            return response()->json(['error'=>"Ce resultat n'a pas encoré été transmis"],401);
+        }else{
+            $resultat->archieved_at = Carbon::now();
+            $resultat->save();
+            return response()->json(['resultat'=>$resultat]);
+        }
+    }
+
+    /**
+     * Passed the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function transmettre($id)
+    {
+        $validation = validatedId($id,$this->table);
+        if(!is_null($validation))
+            return $validation;
+
+        $resultat = ConsultationObstetrique::with(['dossier','consultation'])->find($id);
+        $resultat->passed_at = Carbon::now();
+        $resultat->save();
+
+        return response()->json(['resultat'=>$resultat]);
+
     }
 }
