+++ app/Http/Controllers/Api/ExamenCliniqueController.php
@@ -88,6 +88,12 @@ class ExamenCliniqueController extends Controller
         $validation = validatedId($id,$this->table);
         if(!is_null($validation))
             return $validation;
+
+        $isAuthor = checkIfIsAuthorOrIsAuthorized("ExamenClinique",$id,"create");
+        if($isAuthor->getOriginalContent() == false){
+            return response()->json(['error'=>"Vous ne pouvez modifié un élement que vous n'avez crée"],401);
+        }
+
         ExamenClinique::whereId($id)->update($request->validated());
         $examenClinique = ExamenClinique::find($id);
         return response()->json(['examenClinique'=>$examenClinique]);
@@ -104,6 +110,11 @@ class ExamenCliniqueController extends Controller
         $validation = validatedId($id,$this->table);
         if(!is_null($validation))
             return $validation;
+
+        $isAuthor = checkIfIsAuthorOrIsAuthorized("ExamenClinique",$id,"create");
+        if($isAuthor->getOriginalContent() == false){
+            return response()->json(['error'=>"Vous ne pouvez modifié un élement que vous n'avez crée"],401);
+        }
 
         $examenClinique = ExamenClinique::find($id);
         ExamenClinique::destroy($id);
