+++ app/Http/Controllers/Api/ParametreCommunController.php
@@ -92,6 +92,12 @@ class ParametreCommunController extends Controller
         if(!is_null($validation))
             return $validation;
 
+        $isAuthor = checkIfIsAuthorOrIsAuthorized("ParametreCommun",$id,"create");
+        if($isAuthor->getOriginalContent() == false){
+            return response()->json(['error'=>"Vous ne pouvez modifié un élement que vous n'avez crée"],401);
+        }
+
+
         ParametreCommun::whereId($id)->update($request->validated());
         $parametreCommun = ParametreCommun::with('consultation')->find($id);
 
@@ -115,6 +121,14 @@ class ParametreCommunController extends Controller
         $validation = validatedId($id,$this->table);
         if(!is_null($validation))
             return $validation;
+
+
+        $isAuthor = checkIfIsAuthorOrIsAuthorized("ParametreCommun",$id,"create");
+        if($isAuthor->getOriginalContent() == false){
+            return response()->json(['error'=>"Vous ne pouvez modifié un élement que vous n'avez crée"],401);
+        }
+
+
         $parametreCommun = ParametreCommun::with('consultation')->find($id);
         ParametreCommun::destroy($id);
         return response()->json(['parametreCommun'=>$parametreCommun]);
